Add auto-preview of image as labels are changed
This adds functionality to live-preview translations as they're
changed in the form. To do this, it changes the image API to
be able to retrieve a modified image via a GET request (rather
than POST as it was before) so that to do the reload all we
need to do in the front-end is change the image src.

Then, the API endpoints (there were two: for unmodified and
modified images) are merged into one, and it now optionally
receives new message strings via the URL query string.

The process of adding these strings back into SvgFile was a bit
cumbersome, so a new method SvgFile::setTranslations() was
added, which handles adding the new text and tspan node info.

Tests were updated, but the broken controller test remains for
the time being.

Bug: https://phabricator.wikimedia.org/T207203